support setting max-heap in a jnlp property
The maximum heap memory used by MW can now be set 
using a property element in the jnlp.

<resources>
  <property name="mw.maxHeapSize" value="512m"/>
</resources>

MW starts up in two stages and to affect the memory
used by the second stage it must be passed as a property
instead of as an attribute on the jnlp j2se element.