Re-enable Dan's Run/Stop script in "Oil and Water"
This is somehow interfering with the action of the Run button
script Dan wrote in the text box in the lower model
on this MW page:

http://lab.dev.concord.org/imports/legacy-mw-content/sam-activities/intermolecular-attraction/original-interactives-in-pages/page3/page3.cml

"Run" starts the model. "Stop" stops the model, but 
when "Run" is clicked one more time the model goes forward
one step and stops.